// kadmium-dmx.reactui/src/generated/SetFixtureAttribute.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: SetFixtureAttribute
// ====================================================

export interface SetFixtureAttribute {
  setFixtureAttribute: number;
}

export interface SetFixtureAttributeVariables {
  universeId: number;
  fixture: number;
  attribute: string;
  value: number;
}